Manager statistic done

+++ database/create.sql
@@ -193,8 +193,8 @@ CREATE TABLE "NhuYeuPham"
 CREATE TABLE "ChiTietMuaGoiNYP"(
 	"LichSuMua" int4 NOT NULL,
 	"NhuYeuPham" int4 NOT NULL,
-	"SoLuong" numeric(19,2),
-	"DonGia" numeric(19,0), 
+	"SoLuong" int4,
+	"DonGia" int4, 
 	
 	PRIMARY KEY ("LichSuMua","NhuYeuPham")
